fix up payments
