+++ mifospay/src/main/res/layout/fragment_send.xml
@@ -223,6 +223,8 @@
             android:orientation="horizontal">
 
 
+
+
         </LinearLayout>
 
     </LinearLayout>
